+++ Packages/com.holoi.reality.mofa-the-hunting/Runtime/Scripts/UI/MofaHuntingDragonControllerUIPanel.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 using Holoi.Library.HoloKitApp;
 using Holoi.Library.HoloKitApp.UI;
@@ -5,31 +6,87 @@ using HoloKit;
 
 namespace Holoi.Reality.MOFATheHunting.UI
 {
-    public class MofaHuntingDragonControllerUIPanel : HoloKitAppUIPanel
+    public class MofaHuntingDragonControllerUIPanel : HoloKitAppUIPanel_MonoAR
     {
         public override string UIPanelName => "MofaHuntingDragonController";
 
         public override bool OverlayPreviousPanel => true;
 
+        [Header("MOFA: The Hunting")]
+        [SerializeField] private RectTransform _spawnDragonButton;
+
+        [SerializeField] private RectTransform _monoPanel;
+
+        [SerializeField] private RectTransform _dragonControlPanel;
+
+        private const float SpawnDragonButtonRotationSpeed = 20f;
+
+        public static event Action OnSpawnDragonButtonPressed;
+
+        public static event Action OnDragonPrimaryAttackButtonPressed;
+
+        public static event Action OnDragonSecondaryAttackButtonPressed;
+
+        public static event Action OnDragonLockButtonPressed;
+
         private void Start()
         {
-            HoloKitCamera.Instance.ForceScreenOrientation = false;
+            MofaHuntingRealityManager.OnDragonSpawned += OnDragonSpawned;
+            MofaHuntingRealityManager.OnDragonDied += OnDragonDied;
+
+            _spawnDragonButton.gameObject.SetActive(true);
+            _monoPanel.gameObject.SetActive(true);
+            _dragonControlPanel.gameObject.SetActive(false);
+        }
+
+        private void OnDestroy()
+        {
+            MofaHuntingRealityManager.OnDragonSpawned -= OnDragonSpawned;
+            MofaHuntingRealityManager.OnDragonDied -= OnDragonDied;
         }
 
         private void Update()
         {
             if (Screen.orientation != ScreenOrientation.LandscapeLeft)
                 Screen.orientation = ScreenOrientation.LandscapeLeft;
+
+            // Rotate the spawn dragon button if it is active
+            if (_spawnDragonButton.gameObject.activeSelf)
+            {
+                _spawnDragonButton.Rotate(new Vector3(0f, 0f, 1f), -SpawnDragonButtonRotationSpeed * Time.deltaTime);
+            }
         }
 
-        private void OnDestroy()
+        public void OnSpawnDragonButtonPressedFunc()
         {
-            HoloKitCamera.Instance.ForceScreenOrientation = true;
+            OnSpawnDragonButtonPressed?.Invoke();
         }
 
-        public void OnExitButtonPressed()
+        private void OnDragonSpawned()
         {
-            HoloKitApp.Instance.UIPanelManager.PopUIPanel();
+            _spawnDragonButton.gameObject.SetActive(false);
+            _dragonControlPanel.gameObject.SetActive(true);
+        }
+
+        private void OnDragonDied()
+        {
+            _spawnDragonButton.gameObject.SetActive(true);
+            _dragonControlPanel.gameObject.SetActive(false);
+        }
+
+        public void OnDragonPrimaryAttackButtonPressedFunc()
+        {
+            OnDragonPrimaryAttackButtonPressed?.Invoke();
+        }
+
+        public void OnDragonSecondaryAttackButtonPressedFunc()
+        {
+            OnDragonSecondaryAttackButtonPressed?.Invoke();
+        }
+
+        public void OnDragonLockButtonPressedFunc()
+        {
+            OnDragonLockButtonPressed?.Invoke();
         }
     }
 }